Add 'all books' html page + put css stylesheet in proper directory
The homepage of the website has a navbar that appears on
every page; needs other modifications.

The 'all books' html page is mainly a first test to retrieve the
backend content and dinamically load it into the page itself.

Other minor changes have been made, like moving the css stylesheet
in the proper folder.