Add populate page content mechanism + add several html pages
This commit provides a local try for populate the content of
the books.html page with the server content, mechanism that
will be modified a bit so it can fit the other pages; the
code is device responsive.

Several menu links are made available with their corresponding
pages, that still have to be formatted.

Minor fixes has been made, like the missing html link closing
tag on every li voices of the menu-list in every page.

May the force be with you.