Strengthen the script behind the menu button
The functionality has been indeed improved because,
if a user had previously scrolled the page, the button
was sending back to the top page the user each time it
was pressed again: now the scroll position is saved and
the script will scroll back the page to where the user was.

We are at the fotofinish now.